+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Change log
 
+## 0.19.0
+
+- rename write to writeFlatten
+- rename writeNested to write
+
 ## 0.18.0 1-2-2021
 
 - add ancestor key in ReadWriteCreator and MetaTypes, this key keep all ancestors in order